Improve control center updates.
This commit improves the speed & accuracy of info center now playing updates.

- Instead of passing AudioPlayer.rate (which is a stored state to be used to correct the rate when the player starts playing) we pass `wrapper.rate` which is the current playback rate. This fixes an issue where the now playing timeline was showing glitching when toggling between pause & play, when seeking while paused and a few other situations.
- have updateNowPlayingPlaybackValues update all values in one go
- remove now unused private `AudioPlayer#updateNowPlayingDuration`, `AudioPlayer#updateNowPlayingRate`
- Change `func AVWrapper(seekTo seconds: Int, didFinish: Bool) {` to have `Double` type for seconds
- also update now playing playback values on .ready & .loading events
- in AudioPlayer#load first reset the playback values using the newly introduced `nowPlayingInfoController.setWithoutUpdate`
- NowPlayingInfoController: extract private `update` function from the various `set` methods and avoid `EXC_BAD_ACCESS` when updating `infoCenter.nowPlayingInfo` by making a copy of the info dictionary

+++ Example/Tests/AVPlayerWrapperTests.swift
@@ -215,7 +215,7 @@ class AVPlayerWrapperDelegateHolder: AVPlayerWrapperDelegate {
     
     var stateUpdate: ((_ state: AVPlayerWrapperState) -> Void)?
     var didUpdateDuration: ((_ duration: Double) -> Void)?
-    var didSeekTo: ((_ seconds: Int) -> Void)?
+    var didSeekTo: ((_ seconds: Double) -> Void)?
     var itemDidComplete: (() -> Void)?
     
     func AVWrapper(didChangeState state: AVPlayerWrapperState) {
@@ -230,7 +230,7 @@ class AVPlayerWrapperDelegateHolder: AVPlayerWrapperDelegate {
         
     }
     
-    func AVWrapper(seekTo seconds: Int, didFinish: Bool) {
+    func AVWrapper(seekTo seconds: Double, didFinish: Bool) {
          didSeekTo?(seconds)
     }
