+++ SwiftAudioEx/Classes/AVPlayerWrapper/AVPlayerWrapperDelegate.swift
@@ -14,7 +14,7 @@ protocol AVPlayerWrapperDelegate: AnyObject {
     func AVWrapper(didChangeState state: AVPlayerWrapperState)
     func AVWrapper(secondsElapsed seconds: Double)
     func AVWrapper(failedWithError error: Error?)
-    func AVWrapper(seekTo seconds: Int, didFinish: Bool)
+    func AVWrapper(seekTo seconds: Double, didFinish: Bool)
     func AVWrapper(didUpdateDuration duration: Double)
     func AVWrapper(didReceiveMetadata metadata: [AVTimedMetadataGroup])
     func AVWrapperItemDidPlayToEndTime()
