+++ SwiftAudioEx/Classes/Event.swift
@@ -14,7 +14,7 @@ extension AudioPlayer {
     public typealias PlaybackEndEventData = PlaybackEndedReason
     public typealias SecondElapseEventData = TimeInterval
     public typealias FailEventData = Error?
-    public typealias SeekEventData = (seconds: Int, didFinish: Bool)
+    public typealias SeekEventData = (seconds: Double, didFinish: Bool)
     public typealias UpdateDurationEventData = Double
     public typealias MetadataEventData = [AVTimedMetadataGroup]
     public typealias DidRecreateAVPlayerEventData = ()
